Add a metric shitton of documentation
