+++ ahci/src/main.rs
@@ -75,7 +75,8 @@ use alloc::boxed::Box;
 use alloc::vec::Vec;
 use alloc::sync::Arc;
 use sunrise_libuser::error::{Error, AhciError};
-use sunrise_libuser::ipc::server::{WaitableManager, WorkQueue, port_handler, new_session_wrapper};
+use sunrise_libuser::futures::{WaitableManager, WorkQueue};
+use sunrise_libuser::ipc::server::{port_handler, new_session_wrapper};
 use spin::Mutex;
 use sunrise_libuser::syscalls;
 use sunrise_libuser::ahci::{AhciInterface as IAhciInterface, IDiskProxy, IDisk as _};
