+++ vi/src/main.rs
@@ -36,7 +36,8 @@ use alloc::vec::Vec;
 use alloc::boxed::Box;
 use alloc::sync::{Arc, Weak};
 use crate::libuser::syscalls;
-use crate::libuser::ipc::server::{WaitableManager, port_handler, new_session_wrapper, WorkQueue};
+use crate::libuser::futures::{WaitableManager, WorkQueue};
+use crate::libuser::ipc::server::{port_handler, new_session_wrapper};
 use futures::future::FutureObj;
 use crate::libuser::types::*;
 use spin::Mutex;
